CD-124: [WIP] Caching team membership instead of the team
Currently the class FlowSecurityVoter gets an instance of
AuthenticatorTeamRepository which caches team data, because it has a
caching version of "security.authenticator.client" service.

This change relies on the new TeamMembershipFinder interface. When
caching is disabled, FlowSecurityVoter gets an instance of
AuthenticatorTeamMembershipFinder to retrieve the team membership of the
current user.
When caching is on, then this object is decorated with
CachedAuthenticatorTeamMembershipFinder. This way (team, user) lookup is
done only once during cache lifetime.