CD-114: Add ability to generate documentation for configuration schema
The idea is to define the documentation inside the configuration schema. Schema is represented by a ConfigurationInterface, documentation can be added by the info() and example() method of a config node.
A new console command "app:config:documentation" is provided to generate the documentation for continuous-pipe.yml in Markdown format.