CD-223: Secure BitBucket webhooks
In order to prevent calling the public webhook endpoint by others,
BitBucket appends a JSON Web Token (JWT) to the HTTP request.
This JWT is signed with an algorithm of their choice using the shared
secret key they generate when installing the ContinuousPipe
add-on.
This commit validates each incoming webhook request based on this
using the lcobucci/jwt library.