CD-228: Refactoring: create EarlyAccessToggle on demand
Currently session is used when Symfony DI creates the EarlyAccessToggle service. This cause a problem when building the application container, because "composer install" triggers cache clear which involves building the Symfony DI container.
To avoid that this commit passes the factory to all locations where EarlyAccessToggle is needed. This way EarlyAccessToggle is created only when used and not during DI compilation.