CD-229: Add context to log entries originating from webhook API calls
When code repositories uses webhooks context data can be fetched from the events sent to the message bus.
This commit adds a middleware which intercepts these events and records the context.
In addition to that UserActivityContextProvider and UserActivityContextAggregated is introduced to abstract the source of the context collected from. That way UserActivityContextProcessor does not need to know the data sources, only UserActivityContextAggregated has this information.