CD-239: Add proper error handling for saving team memberships
When constraint violation occurs during saving a team membership, then \Doctrine\ORM\UnitOfWork::commit() method calls the \Doctrine\ORM\EntityManagerInterface::close() method. As a result of this, the next database operation will fail with the following error:
"The EntityManager is closed."

This commit eliminates this error by loading the entity first and modifying that.