CD-583: Make notifications asynchronous
In order to make notifications asynchronous, the original notifier
object is decorated with the TransformCallsToCommand object which
encapsulates the request arguments into a Message object for later
processing by the command bus.

Messages are procesed by the DoNotify class which invokes the original
synchronous notify() method with the arguments extracted from the
message.